v5 - refactor input page

+++ lib/src/domain/models/v3/design/input_data/object_data_design.dart
@@ -12,7 +12,7 @@ class ObjectDataDesign with _$ObjectDataDesign {
     // Адрес объекта
     required final String address,
     // Площадь объекта
-    required final double square,
+    required final int square,
     // Сроки работ
     required final int deadlineValue,
   }) = _ObjectDataDesign;
